Fix i2c transaction bug - uint8 limitation
When attempting to send a payload larger than 256 bytes, sendI2cCommand would fail with a `i2cAddQueue(): realloc Failure`. This was because the loop would run forever causing stack overflow.

+++ src/SparkFun_u-blox_GNSS_Arduino_Library.cpp
@@ -2728,7 +2728,7 @@ sfe_ublox_status_e SFE_UBLOX_GNSS::sendI2cCommand(ubxPacket *outgoingUBX, uint16
   uint16_t startSpot = 0;
   while (bytesToSend > 1)
   {
-    uint8_t len = bytesToSend;
+    uint16_t len = bytesToSend;
     if (len > i2cTransactionSize)
       len = i2cTransactionSize;
